requirements.txt
# =============================================================================
# Desktop Index - Python依存パッケージ
# =============================================================================
# このファイルには、アプリケーションに必要な全てのPythonパッケージを定義します。
#
# インストール方法:
#   pip install -r requirements.txt
# =============================================================================

# ---------------------------------------------------------------------------
# Webフレームワーク
# ---------------------------------------------------------------------------
# FastAPI: 高速な非同期Webフレームワーク
# uvicorn: ASGIサーバー（FastAPIの実行に必要）
# jinja2: HTMLテンプレートエンジン
# python-multipart: フォームデータの処理に必要
# ---------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6

# ---------------------------------------------------------------------------
# 検索エンジンクライアント
# ---------------------------------------------------------------------------
# meilisearch: Meilisearch公式Pythonクライアント
# ---------------------------------------------------------------------------
meilisearch>=0.31.0

# ---------------------------------------------------------------------------
# ドキュメント解析ライブラリ
# ---------------------------------------------------------------------------
# pdfplumber: PDF からテキストを抽出
# python-docx: Word (.docx) ファイルの読み取り
# openpyxl: Excel (.xlsx) ファイルの読み取り
# python-magic: ファイルタイプの自動検出
# chardet: テキストファイルのエンコーディング検出
# ---------------------------------------------------------------------------
pdfplumber>=0.10.3
python-docx>=1.1.0
openpyxl>=3.1.2
python-magic>=0.4.27  # Linux用（libmagic1はDockerfileでインストール済み）
chardet>=5.2.0

# ---------------------------------------------------------------------------
# 設定管理
# ---------------------------------------------------------------------------
# pyyaml: YAML設定ファイルの読み込み
# pydantic: データバリデーション、設定管理
# pydantic-settings: 環境変数からの設定読み込み
# ---------------------------------------------------------------------------
pyyaml>=6.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0

# ---------------------------------------------------------------------------
# 非同期処理・スケジューリング
# ---------------------------------------------------------------------------
# aiofiles: 非同期ファイルI/O
# apscheduler: タスクスケジューラー（定期実行用）
# ---------------------------------------------------------------------------
aiofiles>=23.2.1
apscheduler>=3.10.4

# ---------------------------------------------------------------------------
# ユーティリティ
# ---------------------------------------------------------------------------
# httpx: 非同期HTTPクライアント（Meilisearchとの通信用）
# watchdog: ファイルシステム監視（将来の拡張用）
# ---------------------------------------------------------------------------
httpx>=0.26.0
